backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
PyJWT>=2.8.0
python-multipart>=0.0.6